Feat: working solution with 2 nodes
I overcomplicated the logic and tried to send the whole data set each
time a new element was received, simplified the logic and just send the
data as I get it to the other nodes. As a basic implementation, this
seems to be fine, will test and read up on improving.